fix templates, reorganize, add strict template checking
