+++ go.mod
@@ -4,13 +4,14 @@ go 1.23.0
 
 require (
 	github.com/ddbgio/cookie/v2 v2.0.2
-	github.com/lmittmann/tint v1.0.5
-	github.com/mattn/go-isatty v0.0.20
+	github.com/ddbgio/log v1.0.0
 	github.com/stretchr/testify v1.9.0
 )
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/lmittmann/tint v1.0.5 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	golang.org/x/sys v0.6.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
