+++ static/html/login.html
@@ -1,3 +1,4 @@
+<!DOCTYPE html>
 {{ template "head" . }}
 <html lang="en">
 	<body>
